#1190: implemented requested changes adjusted hierarchy using new IdeaBasedUrlUpdater
